requirements.txt
numpy
pylatex
pyyaml
